requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
mysql-connector-python
numpy